+++ backhand/src/traits/mod.rs
@@ -1,5 +1,10 @@
 //! Shared traits for v3 and v4 SquashFS implementations
 
+pub mod block_reader;
+#[cfg(not(feature = "parallel"))]
+pub mod block_reader_no_parallel;
+#[cfg(feature = "parallel")]
+pub mod block_reader_parallel;
 pub mod compression;
 pub mod error;
 pub mod filesystem;